setting page migrated to view - index model
